+++ r-package/R/transit_network_to_sf.R
@@ -42,6 +42,15 @@ transit_network_to_sf <- function(r5r_core) {
   routes_df[, geometry := sf::st_as_sfc(geometry)]
   routes_sf <- sf::st_sf(routes_df, crs = 4326) # WGS 84
 
+    # supress warning
+    defaultW <- getOption("warn")
+    options(warn = -1)
+
+    # fix eventual invalid geometries
+    if( any(FALSE, sf::st_is_valid(routes_sf)) ){
+      routes_sf <- sf::st_make_valid(routes_sf)
+    }
+
   # Convert stops to SF (point)
   stops_df <- java_to_dt(network$get(1L))
   stops_df[, lat := ifelse(lat==-1,NA,lat)][, lon := ifelse(lon==-1,NA,lon)]
@@ -50,6 +59,9 @@ transit_network_to_sf <- function(r5r_core) {
 
   # gather in a list
   transit_network <- list(stops = stops_sf, routes = routes_sf)
+
+  # restore warnings
+  options(warn = defaultW)
 
   return(transit_network)
 }